Create class User & Film 26.06.2022

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -20,5 +20,5 @@ public class Film {
     private String description;
     private LocalDate releaseDate;
     @Positive
-    private double duration;
+    private int duration;
 }
